+++ src/DateRetrieval.ts
@@ -1,0 +1,85 @@
+import { DateTime } from "luxon";
+import pascua from "pascua";
+import Month from "./Month";
+import WeekDay from "./WeekDay";
+
+export function nextWeekDay(dateTime: DateTime, weekDay: WeekDay): DateTime {
+  let ret = dateTime;
+
+  do {
+    ret = ret.plus( {
+      days: 1,
+    } );
+  } while (ret.weekday !== weekDay);
+
+  return ret;
+}
+
+export function lastWeekDay(dateTime: DateTime, weekDay: WeekDay): DateTime {
+  let ret = dateTime;
+
+  do {
+    ret = ret.minus( {
+      days: 1,
+    } );
+  } while (ret.weekday !== weekDay);
+
+  return ret;
+}
+
+export function nthWeekDayInMonth(
+  year: number,
+  month: Month,
+  weekDay: WeekDay,
+  nth: number,
+  reverse = false,
+): DateTime|undefined {
+  let ret = DateTime.local(year, month);
+
+  if (!ret.isValid)
+    return undefined;
+
+  if (reverse)
+    ret = ret.endOf("month");
+  else
+    ret = ret.startOf("month");
+
+  let count = 0;
+
+  if (ret.weekday === weekDay)
+    count++;
+
+  while (count !== nth) {
+    if (reverse) {
+      ret = ret.minus( {
+        days: 1,
+      } );
+    } else {
+      ret = ret.plus( {
+        days: 1,
+      } );
+    }
+
+    if (ret.month !== month)
+      return undefined;
+
+    if (ret.weekday === weekDay)
+      count++;
+  }
+
+  return ret;
+}
+
+export function mothersDayOf(year: number): DateTime {
+  return <DateTime>nthWeekDayInMonth(year, Month.MAY, WeekDay.SUNDAY, 1);
+}
+
+export function thanksGivingDayOf(year: number): DateTime {
+  return <DateTime>nthWeekDayInMonth(year, Month.NOVEMBER, WeekDay.THURSDAY, 4);
+}
+
+export function easterOf(year: number): DateTime {
+  const { day, month } = pascua(year);
+
+  return DateTime.local(year, month, day);
+}