+++ src/Month.spec.ts
@@ -1,16 +1,16 @@
 import Month from "./Month";
 
 it("month enum", () => {
-  expect(Month.JANUARY).toBe(0);
-  expect(Month.FEBRUARY).toBe(1);
-  expect(Month.MARCH).toBe(2);
-  expect(Month.APRIL).toBe(3);
-  expect(Month.MAY).toBe(4);
-  expect(Month.JUNE).toBe(5);
-  expect(Month.JULY).toBe(6);
-  expect(Month.AUGUST).toBe(7);
-  expect(Month.SEPTEMBER).toBe(8);
-  expect(Month.OCTOBER).toBe(9);
-  expect(Month.NOVEMBER).toBe(10);
-  expect(Month.DECEMBER).toBe(11);
+  expect(Month.JANUARY).toBe(1);
+  expect(Month.FEBRUARY).toBe(2);
+  expect(Month.MARCH).toBe(3);
+  expect(Month.APRIL).toBe(4);
+  expect(Month.MAY).toBe(5);
+  expect(Month.JUNE).toBe(6);
+  expect(Month.JULY).toBe(7);
+  expect(Month.AUGUST).toBe(8);
+  expect(Month.SEPTEMBER).toBe(9);
+  expect(Month.OCTOBER).toBe(10);
+  expect(Month.NOVEMBER).toBe(11);
+  expect(Month.DECEMBER).toBe(12);
 } );