+++ src/TimeUnit.ts
@@ -1,9 +1,11 @@
 enum TimeUnit {
-  MILLISECONDS,
-  SECONDS,
-  MINUTES,
-  HOURS,
-  DAYS,
+  MILLISECONDS = "milliseconds",
+  SECONDS = "seconds",
+  MINUTES = "minutes",
+  HOURS = "hours",
+  DAYS = "days",
+  MONTHS = "months",
+  YEARS = "years",
 }
 
 export default TimeUnit;