+++ src/WeekDayJS.ts
@@ -1,11 +1,11 @@
-enum WeekDay {
+enum WeekDayJS {
   MONDAY = 1,
   TUESDAY = 2,
   WEDNESDAY = 3,
   THURSDAY = 4,
   FRIDAY = 5,
   SATURDAY = 6,
-  SUNDAY = 7
+  SUNDAY = 0
 }
 
-export default WeekDay;
+export default WeekDayJS;