+++ src/index.ts
@@ -1,24 +1,46 @@
 export {
-  between, daysBetween,
+  daysBetween,
   millisBetween,
 } from "./Between";
+
 export {
-  destructDate,
-  destructDateUTC,
+  easterOf,
+  lastWeekDay,
+  mothersDayOf,
+  nextWeekDay,
+  nthWeekDayInMonth,
+  thanksGivingDayOf,
+} from "./DateRetrieval";
+
+export {
+  destructJSDate,
 } from "./DestructDate";
+
 export {
   default as Month,
 } from "./Month";
+
+export {
+  default as MonthJS,
+} from "./MonthJS";
+
 export {
   getTimestamp,
   getTimestampNow,
 } from "./Timestamp";
+
 export {
   default as TimeUnit,
 } from "./TimeUnit";
+
 export {
-  utcToLocal,
-} from "./TimeUtils";
+  isWeekend,
+} from "./Utils";
+
 export {
   default as WeekDay,
 } from "./WeekDay";
+
+export {
+  default as WeekDayJS,
+} from "./WeekDayJS";