+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0
+* Add _moment_ library
+* Remove functions which _moment_ can do directly
+* Add _destructMomentumDate_ function
+* Add some retrieval functions: nextWeekDay, lastWeekDay, nthWeekDayInMonth...
+* Add Calendar
+
 ## 1.0.0
 
 * Initial release